Move Timer and TimerKey definition to restate-types to make shareable
